Clingy: refactoring and general improvements
Add accounting for BUMP|WAIT states, too, since we'll want to know (and
add asserts that we're tracking correctly)

Order revocable spans by JUNK size, rather than absolute size.  Aim for
more bang per buck, as it were.

Enrich revocation policy

Conditionalize many things on __debug__ to improve performance.  Add
some warnings about the interaction of __debug__ and --paranoia

Make bulk of logic into an AbstractBaseClass with placement policy
stubs.  Call stack information is not yet plumbed through, though.
